TT#14008 improve timerthread performance
The only reason the wait time in timer threads is capped at 100 ms is to
catch the daemon's shutdown. Since these threads already use a
condition variable to be woken up, we can use the same condition to wake
up the threads during a shutdown, allowing for a much longer (possibly
unlimited) sleep time in timer threads.

Change-Id: I3b5f280a4d502176a731e8b52b1b731022c8db4f